Soft delete implementation for Locacion and Producto
This commit introduces a soft delete mechanism for the `Locacion` and `Producto` entities.

Instead of physically deleting records, an `activo` (active) boolean field has been added to both models. When a delete operation is performed, this field is set to `false`.

DAOs for `Locacion` and `Producto` have been updated to filter by `activo = 1` in all SELECT queries and to update the `activo` field to `0` during delete operations.

The database version has been incremented to 13.